requirements.txt
customtkinter
pystray
gspread
google-auth-oauthlib
google-auth-httplib2
keyboard
pillow
google-api-python-client
tkinterdnd2
